+++ src/common/js/private.js
@@ -117,16 +117,6 @@ AccountsTools._preferredLabelInitialResult = function( arg, preferred ){
  * @locus Anywhere
  * @returns {Promise} which resolves to the user document
  */
-AccountsTools._userDocByEmail = async function( email, options={} ){
-    check( email, String );
-    check( options, Object );
-    return Meteor.isClient ? Meteor.callAsync( 'AccountsTools.byEmail', email, options ) : AccountsTools.server.byEmail( email, options );
-};
-
-/*
- * @locus Anywhere
- * @returns {Promise} which resolves to the user document
- */
 AccountsTools._userDocById = async function( id ){
     check( id, String );
     return Meteor.isClient ? Meteor.callAsync( 'AccountsTools.byId', id ) : AccountsTools.server.byId( id );