+++ ChangeLog.md
@@ -10,6 +10,7 @@
     - Add a dependency on Meteor accounts-base
     - Define new AccountsTools.cleanupUserDocument() function, bumping candidate release number
     - Reviewed all functions to use async versions and return Promises (todo #3)
+    - New AccountsTools.byEmail() function
 
 ### 2.0.0